Added Location Input Component
