+++ src/components/WeatherToday/WeatherToday.tsx
@@ -18,6 +18,7 @@ export default function WeatherToday() {
           </Link>
           <img
             src={` https://www.weatherbit.io/static/img/icons/${todayWeather.weather.icon}.png`}
+            alt="Weather background"
           />
           <h5>Rain: {todayWeather.precip}mm/hr</h5>
           <h5>
